Added 404

+++ rooms/views.py
@@ -39,6 +39,8 @@ def UpdateView(request, room_id):
         price_per_night = request.data.get("price_per_night")
         room_number = request.data.get("room_number")
         room = Room.objects.get(room_id=room_id)
+        if not room:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         data = {
             "room_id": room_id,
             "hotel_id": hotel_id if hotel_id else str(room.hotel_id),
@@ -61,6 +63,8 @@ def UpdateView(request, room_id):
 def DeleteView(request, room_id):
     try:
         room = Room.objects.get(room_id=room_id)
+        if not room:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         room.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     except Exception as e:
